# Use a profiler to improve linker performance